+++ packages/geoinfo/package.js
@@ -48,8 +48,8 @@ Package.onUse(function(api) {
   api.addFiles([
     'server/Geoapi.js',
     'server/Geoinfo.js',
-    'server/admin.js',
     'server/places.js',
+    'server/admin.js',    
   ],'server');
 
 });